requirements.txt
GitPython
revChatGPT
tiktoken
